// packages/examples-featured/src/components/Register/Register.enums.ts
export enum RegisterDataTestId {
    Email = "email-test-id",
    EmailInvalid = "email-invalid",
    EmailLoading = "email-loading",
    EmailValid = "email-valid",
    Message = "message",
    Pending = "pending",
    Submit = "submit",
    Username = "username-test-id",
    UsernameInvalid = "username-invalid",
    UsernameLoading = "username-loading",
    UsernameValid = "username-valid"
}